requirements.txt
python-dotenv
anthropic>=0.18.0
openai>=2.0.0
google-genai>=1.0.0
pydantic
httpx
streamlit
cryptography
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
mcp>=1.0.0
fastapi
uvicorn
opencv-python 
numpy
tiktoken
markdownify
langfuse
colorama
e2b